Allow --raw argument
This fixes a bug where the `--raw` argument was used in conjunction with `-o json`.
When the `--raw` argument is given we don't need to inject the `-o json`
flag to the kubectl command as that's not allowed by kubectl.

No test added as we mock the kubectl response anyway